requirements.txt
bitsandbytes==0.43.3 
transformers==4.44.2
huggingface_hub==0.24.6
torch==2.4.0
pandas==
joblib==1.4.2
scikit-learn==1.3.2
asyncpg==0.29.0
sqlalchemy==2.0.32
fastapi==0.112.2
langchain_huggingface==0.0.3
langchain==0.2.14
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
quart==0.19.6
httpie==3.2.3
